pyproject.toml
[tool.poetry]
name = "zhenxunflow"
version = "1.0.0"
description = ""
authors = ["AkashiCoin <i@loli.vet>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/zhenxun-org/zhenxunflow"
repository = "https://github.com/zhenxun-org/zhenxunflow"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.3.0"
nonebot-adapter-github = "^0.4.1"
githubkit = "^0.11.7"
pre-commit = "^3.3.2"
jinja2 = "^3.1.2"
pydantic-extra-types = "^2.5.0"

[tool.poetry.group.plugin.dependencies]
click = "^8.1.3"
tzdata = "*"

[tool.poetry.group.dev.dependencies]
nonebug = "^0.3.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.6.1"
pytest-asyncio = "^0.23.5"
pytest-xdist = "^3.3.1"
respx = "^0.21.1"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
  "F",     # Pyflakes
  "W",     # pycodestyle warnings
  "E",     # pycodestyle errors
  "UP",    # pyupgrade
  "ASYNC", # flake8-async
  "C4",    # flake8-comprehensions
  "T10",   # flake8-debugger
  "T20",   # flake8-print
  "PYI",   # flake8-pyi
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
  "RUF",   # Ruff-specific rules
]
ignore = [
  "E402",     # module-import-not-at-top-of-file
  "E501",     # line-too-long
  "UP037",    # quoted-annotation
  "RUF001",   # ambiguous-unicode-character-string
  "RUF002",   # ambiguous-unicode-character-docstring
  "RUF003",   # ambiguous-unicode-character-comment
  "ASYNC230", # blocking-open-call-in-async-function
]

[tool.pyright]
typeCheckingMode = "basic"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.report]
exclude_also = ["if TYPE_CHECKING:", "raise NotImplementedError"]

[tool.coverage.run]
omit = ["src/utils/plugin_test.py"]

[tool.nonebot]
adapters = [{ name = "GitHub", module_name = "nonebot.adapters.github" }]
plugin_dirs = ["src/plugins"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
